refactor: 💡 Move Calendar components to CalendarScreen
