+++ components/header/AppHeader.tsx
@@ -1,13 +1,16 @@
 import { View, Text } from "react-native";
 import { styles } from "../../styles/styles";
 
-export default function AppHeader() {
+interface AppHeaderProps {
+  headerText: string;
+  subText: string;
+}
+
+export default function AppHeader(props: AppHeaderProps) {
   return (
     <View style={styles.headerContainer}>
-      <Text style={styles.headerText}>Plany</Text>
-      <Text style={styles.subText}>
-        Mobile App for tracking events in a calendar
-      </Text>
+      <Text style={styles.headerText}>{props.headerText}</Text>
+      <Text style={styles.subText}>{props.subText}</Text>
     </View>
   );
 }
