+++ components/index.ts
@@ -1,0 +1,3 @@
+import AppHeader from "./header/AppHeader";
+
+export { AppHeader };
